Categories

+++ src/main/scala/FastPayments/db/AccountDb.scala
@@ -10,7 +10,7 @@ object AccountDb {
   class AccountTable(tag: Tag) extends Table[Account](tag, "accounts") {
     val id = column[UUID]("id", O.PrimaryKey)
     val username = column[String]("username")
-    val sum = column[Int]("sum")
+    val sum = column[Float]("sum")
 
     override def * = (id, username, sum) <> ((Account.apply _).tupled, Account.unapply _)
   }
